refactor: moving types, schemas and fetch logic for features in own file
